requirements/jenkins.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=requirements/jenkins.txt requirements/base.txt requirements/dev.txt requirements/jenkins.in
#
-e git+https://github.com/vbabiy/djangorestframework-camel-case.git@master#egg=djangorestframework-camel-case
amqp==2.4.2
astroid==1.6.6            # via pylint
attrs==19.1.0
babel==2.7.0
billiard==3.6.0.0
boltons==19.1.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
click==7.0
coverage==4.5.3
django-appconf==1.0.3
django-axes==4.5.4
django-choices==1.6.2
django-compat==1.0.15
django-debug-toolbar==1.11
django-extra-views==0.12.0
django-hijack-admin==2.1.10
django-hijack==2.1.10
django-ipware==2.1.0
django-jenkins==0.110.0
django-redis==4.10.0
django-rosetta==0.9.3
django-sniplates==0.6.0
django-solo==1.1.3
django==2.2.4
djangorestframework==3.10.2
face==19.1.2
factory-boy==2.11.1
faker==1.0.5
flower==0.9.3
freezegun==0.3.11
gemma-zds-client==0.11.0
glom==19.2.0
idna==2.8
isort==4.3.17
kombu==4.5.0
lazy-object-proxy==1.3.1  # via astroid
mccabe==0.6.1             # via pylint
pep8==1.7.1
pillow==6.0.0
pip-tools==4.0.0
polib==1.1.0
psycopg2==2.8.2
pyjwt==1.7.1
pylint==1.9.4
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2019.1
pyyaml==5.1
raven==6.10.0
redis==3.2.1
requests-mock==1.6.0
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
tblib==1.3.2
text-unidecode==1.2
tornado==5.1.1
urllib3==1.24.2
vine==1.3.0
wrapt==1.11.1             # via astroid
zgw-consumers==0.2.1
